Fix windows build
PR #100 broke the Windows build of go-tpm. This change fixes the build
as well as:

  - Simplifying `poll()` to only take a file, we never use the timeout param
  - Changing file names to allow for testing on non windows/linux platforms
  - Adding build tags to the examples to make it clear they don't build on windows
  - Removing a binary file that accidentally got committed in #104
  - Adds `.gitignore` file to prevent this issue in the future

I don't know how to change the presubmit scripts, but we sould at least make
sure that go-tpm builds on windows (i.e. `GOOS=windows`)
