+++ examples/tpm-keys/tpm-keys.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 // Copyright (c) 2016, Kevin Walsh.  All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
