+++ examples/tpm-sign/common.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package main
 
 import (
