+++ examples/tpm-sign/extend_pcr.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 // Copyright (c) 2018, Ian Haken. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
