+++ examples/tpm-takeownership/own.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 // Copyright (c) 2014, Google LLC All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
