+++ examples/tpm2-seal-unseal/main.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 // Copyright (c) 2018, Google LLC All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
