+++ tpmutil/poll_unix.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"os"
 	"syscall"
-	"time"
 	"unsafe"
 )
 
@@ -16,24 +15,16 @@ type pollFD struct {
 	revents int16
 }
 
-const pollNoTimeout = -1
-
 // poll blocks until the file descriptior is ready for reading or an error occurs.
-func poll(f *os.File, timeout time.Duration) error {
+func poll(f *os.File) error {
 	var (
 		fd = &pollFD{
 			fd:     int32(f.Fd()),
 			events: 0x1, // POLLIN
 		}
 		numFD     = 1
-		timeoutMS int
+		timeoutMS = -1 // Do not set a timeout
 	)
-	// Convert timeout into milliseconds int (or keep -1 for no timeout).
-	if timeout == pollNoTimeout {
-		timeoutMS = pollNoTimeout
-	} else {
-		timeoutMS = int(timeout / time.Millisecond)
-	}
 	_, _, errno := syscall.Syscall(syscall.SYS_POLL, uintptr(unsafe.Pointer(fd)), uintptr(numFD), uintptr(timeoutMS))
 	// Convert errno into an error, otherwise err != nil checks up the stack
 	// will hit unexpectedly on 0 errno.
@@ -43,11 +34,6 @@ func poll(f *os.File, timeout time.Duration) error {
 		return err
 	}
 	// revents is filled in by the kernel.
-	// If nothing happened and we had a timeout set, the there's no data
-	// available to read still.
-	if fd.revents == 0 && timeout > 0 {
-		return fmt.Errorf("poll timed out")
-	}
 	// If the expected event happened, revents should match events.
 	if fd.revents != fd.events {
 		return fmt.Errorf("unexpected poll revents 0x%x", fd.revents)
